// src/app/pucks/create-dataset/rsrcType.interface.ts
export interface  RsrcTypeElement {
  rsrcTypeKey: number;
  rsrcTypeNam: string;
  rsrcValueTypeKey: number;
  rsrcValueTypeNam: string;
  recActiveFlg: boolean;
  resourceTypeUiText: string;
  rqrdValueFlg: boolean;
  resourceValueTypeUiText: string;
  rsrcTypeGroupNam: string;
  resourceTypeGroupUiText: string;
  moduleName: string;
  multiRsrcFlg: boolean;
  rsrcVrsnId: string;
  moduleCtgryNam: string;
  displayOrder: number;
  ruleTypeName: string;
  ruleText: string;
  errorMessage: string;
  inputValueTypeNam: string;
  rsrcTypeValue?: string
  label?: string
  uiText?: any
  data?: any
  value?: any;
}